rpiserver/pyproject.toml
[tool.poetry]
name = "rpiserver"
version = "0.1.0"
description = ""
authors = ["Hannah Willoughby <hannah@hannahwilloughby.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {version = "^0.104.1", extras = ["all"]}
rpi-gpio = "^0.7.1"
jinja2 = "^3.1.2"
requests = "^2.31.0"
pywizlight = "^0.5.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
